Rename mesa-x11 platform to just x11
