+++ debian/control
@@ -266,19 +266,19 @@ Description: Display server for Ubuntu - shared library
  Contains the shared libraries required for the Mir server and client.
 
 # Longer-term these drivers should move out-of-tree
-Package: mir-platform-graphics-mesa-x16
+Package: mir-platform-graphics-x16
 Section: libs
 Architecture: linux-any
 Multi-Arch: same
 Pre-Depends: ${misc:Pre-Depends}
 Depends: ${misc:Depends},
          ${shlibs:Depends},
-Description: Display server for Ubuntu - platform library for X11 Mesa
+Description: Display server for Ubuntu - platform library for X11
  Mir is a display server running on linux systems, with a focus on efficiency,
  robust operation and a well-defined driver model.
  .
  Contains the shared libraries required for the Mir server to interact with
- the X11 platform using the Mesa drivers.
+ the X11 platform.
 
 Package: mir-platform-graphics-mesa-kms16
 Section: libs
