--- debian/mir-platform-graphics-mesa-x16.install
@@ -1,1 +1,0 @@
-usr/lib/*/mir/server-platform/server-mesa-x11.so.16
