+++ debian/mir-platform-graphics-x16.install
@@ -1,0 +1,1 @@
+usr/lib/*/mir/server-platform/server-x11.so.16
