+++ src/platforms/CMakeLists.txt
@@ -68,7 +68,7 @@ if (MIR_BUILD_PLATFORM_MESA_KMS)
     add_subdirectory(mesa/)
 endif()
 
-if (MIR_BUILD_PLATFORM_MESA_X11)
+if (MIR_BUILD_PLATFORM_X11)
     add_subdirectory(x11/)
 endif()
