+++ tests/mir_test_doubles/CMakeLists.txt
@@ -52,12 +52,11 @@ set(
   ${CMAKE_CURRENT_SOURCE_DIR}/mock_drm.cpp
 )
 
-if (MIR_BUILD_PLATFORM_MESA_X11)
+if (MIR_BUILD_PLATFORM_X11)
   include_directories(
     ${PROJECT_SOURCE_DIR}/src/platforms/mesa/server
   )
   list(APPEND MIR_TEST_DOUBLES_PLATFORM_SRCS
-    ${CMAKE_CURRENT_SOURCE_DIR}/mock_gbm.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/mock_x11.cpp
   )
 endif()
