+++ tests/unit-tests/CMakeLists.txt
@@ -12,8 +12,8 @@ if (MIR_BUILD_PLATFORM_MESA_KMS)
   add_definitions(-DMIR_BUILD_PLATFORM_MESA_KMS)
 endif()
 
-if (MIR_BUILD_PLATFORM_MESA_X11)
-  add_definitions(-DMIR_BUILD_PLATFORM_MESA_X11)
+if (MIR_BUILD_PLATFORM_X11)
+  add_definitions(-DMIR_BUILD_PLATFORM_X11)
 endif()
 
 if (MIR_BUILD_PLATFORM_EGLSTREAM_KMS)
@@ -175,13 +175,13 @@ target_link_libraries(
   ${CMAKE_THREAD_LIBS_INIT} # Link in pthread.
 )
 
-if (MIR_BUILD_PLATFORM_MESA_KMS OR MIR_BUILD_PLATFORM_MESA_X11 OR MIR_BUILD_PLATFORM_EGLSTREAM_KMS)
+if (MIR_BUILD_PLATFORM_MESA_KMS OR MIR_BUILD_PLATFORM_EGLSTREAM_KMS)
   target_link_libraries(
     mir_unit_tests
 
     mirsharedmesa-static
   )
-endif (MIR_BUILD_PLATFORM_MESA_KMS OR MIR_BUILD_PLATFORM_MESA_X11 OR MIR_BUILD_PLATFORM_EGLSTREAM_KMS)
+endif (MIR_BUILD_PLATFORM_MESA_KMS OR MIR_BUILD_PLATFORM_EGLSTREAM_KMS)
 
 mir_add_wrapped_executable(mir_umock_unit_tests NOINSTALL
   ${UMOCK_UNIT_TEST_SOURCES}
