+++ tests/unit-tests/platforms/x11/test_graphics_platform.cpp
@@ -80,4 +80,4 @@ public:
     ::testing::NiceMock<mtd::MockX11> mock_x11;
 };
 
-#include "../../test_graphics_platform.h"
+#include "../test_graphics_platform.h"
